Revert back to single file component.
After a long weekend of testing I prefer to have a single file component as it was before the merge. This is because I've a roadmap for the v1.0 that includes a separation of the actual component into two sub components, one for the gallery and another for the theater.