Major update to carousel vue application
This update includes a new revamped carousel system that has been splitted in two different components.
The first component is `PhotoGallery` which is the container of the lists of photos from any source.
The second component is the `Theater` which has the job of displaying the current selected photo from the photo gallery.

Why creating two components? Because one big `Carousel` component was a bit of a mess to maintain and scale, also with the preivous version you have to use both in the same page structure but now they are separated and you can use both as two siblings.

How to they communicate? Simple they relay on an external EventDispatcher that wraps a simple vue instance in order to use `$on` end `$emit` events!

The code is still far from the perfection.

Patch notes:

- Fix index.js in order to use the new version
- Add Theater.vue Component
- Add PhotoGallery.vue component
- Fix carousel.js entry point in order to use the new system
- Remove Carousel.vue component
- Add shared EventDispatcher class