Carousel major update for v0.3
This major update includes a bunch of new features as
well as a new redesigned system for storing the current
state of the carousel!

Introducing: Vuex.

With vuex we are now ready to scale this carousel without cluttering
the main components. In fact vuex can be used to store all the current photos
and the selected one, is that cool?

Our components have less code and they don't have to bother anymore with
a complicated share event handler.

We have also done some minor optimization to the infrastructure removing
useless code and refactoring the current one to be more modular.

Patch notes:
- Add a new folder that contains the store system built with Vuex
- Remove the old shared events handler
- Improve the quality of the components with a complete refactor
- Add components name
- Add Type on components properties
- Remove useless spacing inside the files