Optimization update for v0.4
This update brings a lot of optimizations needed in order to create a scalable carousel for the future and thus we need to bump the version up because it's a major change.

First thing first we tried to set a standard for the code itself with good identation across all files.

We decouple the Photos component by splitting the component in two entities:

- Photos is a general container and it should not be responsible of the Photo itself and his behaviours
- Photo is a single photo component which is  responsible of controlling the single photo loaded from the source.

We also optimize the theater by removing the index as parameter because we can fetch it from the vuex itself, we don't need to pass index as a separate parameter anymore.

We have sone some small changes on the Vuex store that introduce some new getters and an optimized `changeSelectedPhoto` mutation.