Improve the way theater changes photos
This patch brings a new and improved way to navigate between photos.

Previously each time you reach the end or the beginning of the gallery the Theater just vanish  automatically and this wasn't intented at all!

What I want instead is that the gallery never close unless the user clicks on the "x" button or press escape or whatever.

In order to accomplish this I implemented two new mutations and remove the "change index logic" from the theater itself which should not be aware of the index change anyway.

I also fix the undefined error on "changeSelectedPhoto" finally!