Carousel minor update for v0.4.3
This update brings some minor improvements to the
source code.

First I removed the HTTP request from the Photos component
because it does not belongs to it, instead I've created "actions"
inside the Vuex store, which is async and now the component can
 call the action in order to load the photos using the HTTP get request.

This allows everyone to have a source-agnostic gallery. To prove that
I changed the HTTP library from vue-resource to Axios, which is better in my
opinion.

I've also improved the README.

Patch notes:
- Add new action file for the Vuex store.
- Add new "hot to install" section inside readme
- Add new "Contributing guidelines" section inside readme
- Bump the version to v0.4.3
- Add new Axios library for HTTP calls