Carousel major update for v0.5.0
This patch brings a lot of new functionality
as well as a new import system that helps
to easily import the right components as
well as the store system inside your Vue application.

A new set of components have been implemented for
the so awaited comments section!
Yes you can attach comments on each photo and display
them inside the theater. At the moment the comment section
cannot be edited, which means that the comments are in read-only
mode. In the future this will be possible.

I've also improved the readme section as well as the index.html and
I've separated the example from the source file for an easy import inside
your application.

The road to v1.0.0 is far but we have just marked the v0.5.0 which is half of that

Patch notes:
- Add the v0.5.0 to Package.json
- Change dist name from "carousel.js" to "example.js" in order to separate example from the source
- Remove vue instance from carousel.js
- Improve import system for distributing the package with npmjs
- Add new Comment component
- Add new Comments component
- Improve Photo component
- Improve Theater component DOM tree
- Improve Theater component flexbox system
- Add comments component to Theater component
- Add new moment.js dependency
- Improve README and general documentation