Add keyboard navigation 🎂🍾🎆
This patch brings a well requested feature: Keyboard navigation!

To be fair this is the first attempt to it, basically the theater now has a new keyboard event listener attached to it when it's first created. This allow us to intercept every key stroke (on keyup to be precise) and check whenever the user press the right arrow, left arrow or the escape button.